Cleanup core layout, create NavDrawer comp
