+++ src/lib/_documentation/README.md
@@ -1,0 +1,3 @@
+# Documentation Components
+
+This section exists to house components that exist purely for documentation. These are not slated to be released as part of the public library set.
