+++ src/lib/_documentation/ThemeGenerator/Swatches.svelte
