+++ src/lib/_documentation/ThemeGenerator/colors.ts
